Refactor rename questionnaire_onboarding -> onboarding
