+++ shared/src/commonMain/kotlin/org/hyperskill/app/core/injection/AppGraph.kt
@@ -50,7 +50,7 @@ import org.hyperskill.app.project_selection.details.injection.ProjectSelectionDe
 import org.hyperskill.app.project_selection.list.injection.ProjectSelectionListComponent
 import org.hyperskill.app.projects.injection.ProjectsDataComponent
 import org.hyperskill.app.providers.injection.ProvidersDataComponent
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.injection.QuestionnaireOnboardingComponent
+import org.hyperskill.app.users_questionnaire.onboarding.injection.QuestionnaireOnboardingComponent
 import org.hyperskill.app.reactions.injection.ReactionsDataComponent
 import org.hyperskill.app.request_review.injection.RequestReviewDataComponent
 import org.hyperskill.app.request_review.modal.injection.RequestReviewModalComponent
