+++ shared/src/commonMain/kotlin/org/hyperskill/app/core/injection/BaseAppGraph.kt
@@ -94,8 +94,8 @@ import org.hyperskill.app.projects.injection.ProjectsDataComponent
 import org.hyperskill.app.projects.injection.ProjectsDataComponentImpl
 import org.hyperskill.app.providers.injection.ProvidersDataComponent
 import org.hyperskill.app.providers.injection.ProvidersDataComponentImpl
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.injection.QuestionnaireOnboardingComponent
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.injection.QuestionnaireOnboardingComponentImpl
+import org.hyperskill.app.users_questionnaire.onboarding.injection.QuestionnaireOnboardingComponent
+import org.hyperskill.app.users_questionnaire.onboarding.injection.QuestionnaireOnboardingComponentImpl
 import org.hyperskill.app.reactions.injection.ReactionsDataComponent
 import org.hyperskill.app.reactions.injection.ReactionsDataComponentImpl
 import org.hyperskill.app.request_review.injection.RequestReviewDataComponent
