+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/onboarding/domain/analytic/QuestionnaireOnboardingClickedSendHyperskillAnalyticEvent.kt
@@ -1,13 +1,14 @@
-package org.hyperskill.app.users_questionnaire.questionnaire_onboarding.domain.analytic
+package org.hyperskill.app.users_questionnaire.onboarding.domain.analytic
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticPart
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRoute
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
+import ru.nobird.app.core.model.mapOfNotNull
 
 /**
- * Represents click on the choice analytic event.
+ * Represents click on the "Send" button analytic event.
  *
  * JSON payload:
  * ```
@@ -15,27 +16,32 @@ import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTar
  *     "route": "/onboarding/questionnaire",
  *     "action": "click",
  *     "part": "main",
- *     "target": "choice",
+ *     "target": "send",
  *     "context":
  *     {
- *         "source": "Google Play"
+ *         "source": "Google Play",
+ *         "input": "Some text"
  *     }
  * }
  * ```
  *
  * @see HyperskillAnalyticEvent
  */
-class QuestionnaireOnboardingClickedChoiceHyperskillAnalyticEvent(
-    private val selectedChoice: String
+class QuestionnaireOnboardingClickedSendHyperskillAnalyticEvent(
+    private val selectedChoice: String,
+    private val textInputValue: String?
 ) : HyperskillAnalyticEvent(
     HyperskillAnalyticRoute.Onboarding.Questionnaire,
     HyperskillAnalyticAction.CLICK,
     HyperskillAnalyticPart.MAIN,
-    HyperskillAnalyticTarget.CHOICE
+    HyperskillAnalyticTarget.SEND
 ) {
     override val params: Map<String, Any>
         get() = super.params +
             mapOf(
-                PARAM_CONTEXT to mapOf(QuestionnaireOnboardingAnalyticParams.PARAM_SOURCE to selectedChoice)
+                PARAM_CONTEXT to mapOfNotNull(
+                    QuestionnaireOnboardingAnalyticParams.PARAM_SOURCE to selectedChoice,
+                    QuestionnaireOnboardingAnalyticParams.PARAM_INPUT to textInputValue
+                )
             )
 }
