+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/onboarding/domain/analytic/QuestionnaireOnboardingClickedSkipHyperskillAnalyticEvent.kt
@@ -1,0 +1,29 @@
+package org.hyperskill.app.users_questionnaire.onboarding.domain.analytic
+
+import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
+import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
+import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticPart
+import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRoute
+import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
+
+/**
+ * Represents click on the "Skip" button analytic event.
+ *
+ * JSON payload:
+ * ```
+ * {
+ *     "route": "/onboarding/questionnaire",
+ *     "action": "click",
+ *     "part": "main",
+ *     "target": "skip"
+ * }
+ * ```
+ *
+ * @see HyperskillAnalyticEvent
+ */
+object QuestionnaireOnboardingClickedSkipHyperskillAnalyticEvent : HyperskillAnalyticEvent(
+    HyperskillAnalyticRoute.Onboarding.Questionnaire,
+    HyperskillAnalyticAction.CLICK,
+    HyperskillAnalyticPart.MAIN,
+    HyperskillAnalyticTarget.SKIP
+)
