+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/onboarding/domain/analytic/QuestionnaireOnboardingViewedHyperskillAnalyticEvent.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.users_questionnaire.questionnaire_onboarding.domain.analytic
+package org.hyperskill.app.users_questionnaire.onboarding.domain.analytic
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
