+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/onboarding/presentation/QuestionnaireOnboardingReducer.kt
@@ -1,15 +1,15 @@
-package org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation
+package org.hyperskill.app.users_questionnaire.onboarding.presentation
 
 import org.hyperskill.app.SharedResources
 import org.hyperskill.app.core.view.mapper.ResourceProvider
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.domain.analytic.QuestionnaireOnboardingClickedChoiceHyperskillAnalyticEvent
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.domain.analytic.QuestionnaireOnboardingClickedSendHyperskillAnalyticEvent
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.domain.analytic.QuestionnaireOnboardingClickedSkipHyperskillAnalyticEvent
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.domain.analytic.QuestionnaireOnboardingViewedHyperskillAnalyticEvent
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Action
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.InternalAction
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Message
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.State
+import org.hyperskill.app.users_questionnaire.onboarding.domain.analytic.QuestionnaireOnboardingClickedChoiceHyperskillAnalyticEvent
+import org.hyperskill.app.users_questionnaire.onboarding.domain.analytic.QuestionnaireOnboardingClickedSendHyperskillAnalyticEvent
+import org.hyperskill.app.users_questionnaire.onboarding.domain.analytic.QuestionnaireOnboardingClickedSkipHyperskillAnalyticEvent
+import org.hyperskill.app.users_questionnaire.onboarding.domain.analytic.QuestionnaireOnboardingViewedHyperskillAnalyticEvent
+import org.hyperskill.app.users_questionnaire.onboarding.presentation.QuestionnaireOnboardingFeature.Action
+import org.hyperskill.app.users_questionnaire.onboarding.presentation.QuestionnaireOnboardingFeature.InternalAction
+import org.hyperskill.app.users_questionnaire.onboarding.presentation.QuestionnaireOnboardingFeature.Message
+import org.hyperskill.app.users_questionnaire.onboarding.presentation.QuestionnaireOnboardingFeature.State
 import ru.nobird.app.presentation.redux.reducer.StateReducer
 
 private typealias ReducerResult = Pair<State, Set<Action>>
