--- shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/questionnaire_onboarding/injection/QuestionnaireOnboardingComponent.kt
@@ -1,10 +1,0 @@
-package org.hyperskill.app.users_questionnaire.questionnaire_onboarding.injection
-
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Action
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Message
-import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.ViewState
-import ru.nobird.app.presentation.redux.feature.Feature
-
-interface QuestionnaireOnboardingComponent {
-    val questionnaireOnboardingFeature: Feature<ViewState, Message, Action>
-}
